backend/requirements.txt
controlflow>=0.1.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
requests>=2.31.0
openai>=1.35.0
python-dotenv>=1.0.0
pydantic>=2.5.0
python-multipart>=0.0.6
google-auth>=2.23.4
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.108.0
cryptography>=41.0.7
